% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zuericolors.R
\name{get_zuericolors}
\alias{get_zuericolors}
\title{Get Zuericolors Palettes}
\usage{
get_zuericolors(palette, nth)
}
\arguments{
\item{palette}{Name of a specific palette}

\item{nth}{Position of a specific color in a palette, optional}
}
\value{
the color(s) as hex strings
}
\description{
Function to access color values of a specific zuericolors palette.
}
\examples{
# Get colors of palette 'qual12'
get_zuericolors("qual12")

# Get third color value of palette 'qual12'
get_zuericolors("qual12", 3)

# Get first three color values of palette 'qual12'
get_zuericolors("qual12", 1:3)

}
