% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Zuericolors Palettes}
\format{
An object of class \code{list} of length 15.
}
\usage{
palettes
}
\description{
Available color palettes that correspond to the corporate design of the city of Zurich.
}
\keyword{internal}
