% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_zuericolors.R
\name{view_zuericolors}
\alias{view_zuericolors}
\title{View Zuericolors Palettes}
\usage{
view_zuericolors(palette)
}
\arguments{
\item{palette}{Name of a specific palette}
}
\value{

}
\description{
Function to visually inspect values and colors of a specific zuericolors palette
}
\examples{
view_zuericolors("harmonic6")
}
