% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_zuericolors.R
\name{view_zuericolors}
\alias{view_zuericolors}
\title{View Zuericolors Palettes}
\usage{
view_zuericolors(palette)
}
\arguments{
\item{palette}{Name of a specific palette. if it is not provided, all palettes are plotted.}
}
\value{
returns a plot of the requested color/palette
}
\description{
Function to visually inspect values and colors of a specific zuericolors palette
}
\examples{
\dontrun{
view_zuericolors("harmonic6")
}
}
